requirements.txt
Flask==3.0.2
grpcio==1.54.0
python-decouple==3.8
protobuf==4.22.3
